Poprawka awatarowego endpointu
Zastąpiono metodę na PATCH, naprawiono błąd dostępu
